[refactor] 패키지 나누기
