+++ src/test/java/com/jaeyeonling/ladder/domain/PositiveTest.java
@@ -1,5 +1,6 @@
-package com.jaeyeonling.ladder;
+package com.jaeyeonling.ladder.domain;
 
+import com.jaeyeonling.ladder.domain.Positive;
 import com.jaeyeonling.ladder.exception.ShorterThanMinPositiveException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
