+++ src/test/java/com/jaeyeonling/ladder/domain/line/LinesTest.java
@@ -1,5 +1,8 @@
-package com.jaeyeonling.ladder;
+package com.jaeyeonling.ladder.domain.line;
 
+import com.jaeyeonling.ladder.domain.Fixture;
+import com.jaeyeonling.ladder.domain.line.Line;
+import com.jaeyeonling.ladder.domain.line.Lines;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
