+++ src/test/java/com/jaeyeonling/ladder/domain/line/PointGenerateStrategyTest.java
@@ -1,5 +1,6 @@
-package com.jaeyeonling.ladder;
+package com.jaeyeonling.ladder.domain.line;
 
+import com.jaeyeonling.ladder.domain.line.PointGenerateStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
