+++ src/test/java/com/jaeyeonling/ladder/view/format/UserFormatterTest.java
@@ -1,7 +1,7 @@
 package com.jaeyeonling.ladder.view.format;
 
-import com.jaeyeonling.ladder.User;
-import com.jaeyeonling.ladder.Username;
+import com.jaeyeonling.ladder.domain.user.User;
+import com.jaeyeonling.ladder.domain.user.Username;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
