+++ src/test/java/com/jaeyeonling/ladder/view/format/UsernameFormatterTest.java
@@ -1,6 +1,6 @@
 package com.jaeyeonling.ladder.view.format;
 
-import com.jaeyeonling.ladder.Username;
+import com.jaeyeonling.ladder.domain.user.Username;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
